requirements.txt
psycopg2-binary
python-dotenv
PyGithub
requests
pandas
dotenv